Finished implementing parameter values and ranges.
Ilkka Rinne clarified how identifying parameters by name is
intended to work via the scheme attribute even though it
is not required to be a data member for a value or range.
The code will not search for fmi-scheme, and use the parameter
number given. If there is no fmi-scheme, the parameter will
be marked as unknown by setting the parameter number to -1.