Implemented rendering of precipitation areas.
The definition of precipitation areas is strange. The areas
have an exterior, which is fine, plus a list of holes, which
is also fine. However, in addition the area has inner areas
which are recursively defined to be similar precipitation
areas. Thus the area may have an inner area, which may have holes,
and an inner area which may have further holes ad infinitum. It is
not easy to see how one can easily create and edit such constructions.
The WOML spec should be simplified so that infinite recursion is not
possible. Currently Mirwa supports only editing the exterior.

Currently frontier will only render the exterior plus holes, and
ignores the inner precipitation areas.

The actual solution of the linear system of equations for a normal
natural spline to Bezier conversion is triangular, which makes it
easy to solve. However, if you use periodic boundary conditions,
the system becomes slightly non-triangular, and solving the problem
would require a proper Gaussian elimination algorithm, possibly
with proper pivoting to prevent rounding errors.

A much simpler solution is to use pseudo-periodic boundary conditions.
We simply insert to the front the last points of the path, and append
to the end the first points of the path, and after the conversion we
use only the original inner part of the path. This makes the path periodic
enough for practical purposes. Since the boundary conditions require
continuity of the first and second derivatives, we duplicate two
points from each end of the Bezier curve.